requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
exceptiongroup==1.1.3
idna==3.4
influxdb==5.3.1
iniconfig==2.0.0
msgpack==1.0.7
packaging==23.2
paho-mqtt==1.6.1
pluggy==1.3.0
py==1.11.0
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tomli==2.0.1
urllib3==2.0.6
